requirements.txt
Django
django-filter
djangorestframework
pillow
drf_writable_nested
drf-yasg
django-ckeditor